tema3/drive/requirements.txt
Flask==0.12.2
gunicorn==19.7.1
requests==2.18.4
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client